refactor: useWindowSize hook으로 로직 변경
