+++ FE/src/components/interviewPage/InterviewModal/InterviewFinishModal.tsx
@@ -8,7 +8,7 @@ import { PATH } from '@constants/path';
 import { useQueryClient } from '@tanstack/react-query';
 import { QUERY_KEY } from '@constants/queryKey';
 import Confetti from 'react-confetti';
-import { useEffect, useState } from 'react';
+import useWindowSize from '@hooks/pages/Interview/useWindowSize';
 
 type InterviewFinishModalProps = {
   isOpen: boolean;
@@ -20,43 +20,12 @@ const InterviewFinishModal: React.FC<InterviewFinishModalProps> = ({
   closeModal,
 }) => {
   const navigate = useNavigate();
-
   const isLogin = useQueryClient().getQueryState(QUERY_KEY.MEMBER);
-
-  const [windowSize, setWindowSize] = useState({
-    width: window.innerWidth,
-    height: window.innerHeight,
-  });
-
-  useEffect(() => {
-    let timeoutId: number | undefined;
-
-    const handleResize = () => {
-      window.clearTimeout(timeoutId);
-
-      timeoutId = window.setTimeout(() => {
-        console.log({
-          width: window.innerWidth,
-          height: window.innerHeight,
-        });
-        setWindowSize({
-          width: window.innerWidth,
-          height: window.innerHeight,
-        });
-      }, 250);
-    };
-
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-      if (timeoutId !== undefined) window.clearTimeout(timeoutId);
-    };
-  }, []);
+  const windowSize = useWindowSize();
 
   return (
     <>
-      <Modal isOpen={isOpen} closeModal={closeModal}>
+      <Modal
         {isOpen && (
           <Confetti width={windowSize.width} height={windowSize.height} />
         )}
